add message and change UI

+++ lib/widgets/todo_item.dart
@@ -11,7 +11,6 @@ class ToDoItem extends StatelessWidget {
   final onDeleteItem;
   final onUpdateItem;
 
-
   const ToDoItem({
     super.key,
     required this.todo,
@@ -30,31 +29,25 @@ class ToDoItem extends StatelessWidget {
         ),
 
         onTap: () {
-
           Navigator.push(
             context,
             MaterialPageRoute(
-                builder: (context) =>
-                    UpdateTodoPage(
-                      currentText: todo.todoText!,
-                      currentNote: todo.todoNote ,
-                      currentDate: todo.date,
-                      currentTime: todo.time,
-                      
-                      ),
-                    ),
-                     
-                    
+              builder: (context) => UpdateTodoPage(
+                currentText: todo.todoText!,
+                currentNote: todo.todoNote,
+                currentDate: todo.date,
+                currentTime: todo.time,
+              ),
+            ),
           ).then((updatedText) {
             if (updatedText['todoText'] != null && updatedText is Map) {
               onUpdateItem(
-              updatedText['todoText'],
-              updatedText['todoNote'],
-              updatedText['date'],
-              updatedText['time'],
-              todo.id,
+                updatedText['todoText'],
+                updatedText['todoNote'],
+                updatedText['date'],
+                updatedText['time'],
+                todo.id,
               );
-              
             }
           });
         },
@@ -75,8 +68,6 @@ class ToDoItem extends StatelessWidget {
           },
         ),
 
-     
-
         title: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
@@ -89,19 +80,17 @@ class ToDoItem extends StatelessWidget {
               ),
             ),
             const SizedBox(height: 5),
-            
+
             // Text(
             todo.todoNote == null || todo.todoNote!.isEmpty
-            ? const SizedBox.shrink()
-            : Text(
-                todo.todoNote!,
-                style: const TextStyle(
-                  fontSize: 12,
-                  color: tdGrey,
-                ),
-              ),
-
-
+                ? const SizedBox.shrink()
+                : Text(
+                    todo.todoNote!,
+                    style: const TextStyle(
+                      fontSize: 12,
+                      color: tdGrey,
+                    ),
+                  ),
           ],
         ),
 
@@ -112,20 +101,21 @@ class ToDoItem extends StatelessWidget {
           child: Column(
             children: [
               todo.time == null
-                ? const SizedBox.shrink() // hides the widget, takes no space
-                : Text(
-                    todo.time!.format(context),
-                    style: const TextStyle(fontSize: 10, color: tdGrey),
-                  ),
+                  ? const SizedBox.shrink() // hides the widget, takes no space
+                  : Text(
+                      todo.time!.format(context),
+                      style: const TextStyle(fontSize: 10, color: tdGrey),
+                    ),
 
               const SizedBox(height: 2),
 
-         // ✅ date
-              todo.date == null? const SizedBox.shrink():
-              Text(
-              '${todo.date!.year}-${todo.date!.month.toString().padLeft(2, '0')}-${todo.date!.day.toString().padLeft(2, '0')}',
-                style: const TextStyle(fontSize: 10, color:tdGrey),
-              )
+              // ✅ date
+              todo.date == null
+                  ? const SizedBox.shrink()
+                  : Text(
+                      '${todo.date!.year}-${todo.date!.month.toString().padLeft(2, '0')}-${todo.date!.day.toString().padLeft(2, '0')}',
+                      style: const TextStyle(fontSize: 10, color: tdGrey),
+                    )
             ],
           ),
         ),
@@ -134,8 +124,31 @@ class ToDoItem extends StatelessWidget {
         trailing: IconButton(
           icon: const Icon(Icons.delete, color: tdRed),
           onPressed: () {
+            showDialog(context: context,
+             builder: (context) {
+              return AlertDialog(
+                title: const Text('Delete Todo'),
+                content: const Text('Are you sure you want to delete this todo?'),
+                actions: [
+                  TextButton(
+                    onPressed: () {
+                      Navigator.of(context).pop(); // Close the dialog
+                    },
+                    child: const Text('Cancel'),
+                  ),
+                  TextButton(
+                    onPressed: () {
+                      onDeleteItem(todo.id);
+                      Navigator.of(context).pop(); // Close the dialog
+                    },
+                    child: const Text('Delete',style: TextStyle(color: tdRed)),
+                  ),
+                ],
+              );
+             });
+
             // print('click delet button');
-            onDeleteItem(todo.id);
+            // onDeleteItem(todo.id);
           },
         ),
       ),
